+++ src/discljord/messaging/impl.clj
@@ -55,13 +55,44 @@
                     nil))
     response))
 
+(defmethod dispatch-http :create-guild-ban
+  [process endpoint [prom user-id
+                     & {:keys [user-agent delete-message-days reason]}]]
+  (let [guild-id (-> endpoint
+                     ::ms/major-variable
+                     ::ms/major-variable-value)
+        query {}
+        query (if delete-message-days
+               (assoc query :delete-message-days delete-message-days)
+               query)
+        query (if reason
+               (assoc query :reason reason)
+               query)
+        response @(http/put (api-url (str "/guilds/" guild-id "/bans/" user-id))
+                            {:headers (auth-headers (::ds/token @process) user-agent)
+                             :query-params query})
+        success? (= 204 (:status response))]
+    (deliver prom success?)
+    response))
+
 (defmethod dispatch-http :get-guild-roles
-  [process endpoint [prom guild-id & {:keys [user-agent]}]]
-  (let [channel (-> endpoint
+  [process endpoint [prom & {:keys [user-agent]}]]
+  (let [guild-id (-> endpoint
                     ::ms/major-variable
                     ::ms/major-variable-value)
         response @(http/get (api-url (str "/guilds/" guild-id "/roles"))
                             {:headers (auth-headers (::ds/token @process) user-agent)})
+        json-msg (json/read-str (:body response))]
+    (deliver prom (if json-msg
+                    (clean-json-input json-msg)
+                    nil))
+    response))
+
+(defmethod dispatch-http :create-dm
+  [process endpoint [prom user-id & {:keys [user-agent]}]]
+  (let [response @(http/post (api-url "/users/@me/channels")
+                             {:headers (auth-headers (::ds/token @process) user-agent)
+                              :body (json/write-str {:recipient_id user-id})})
         json-msg (json/read-str (:body response))]
     (deliver prom (if json-msg
                     (clean-json-input json-msg)