+++ CHANGELOG.md
@@ -9,10 +9,14 @@ Discljord follows semantic versioning.
  - Request :guild-members-chunk events
  - Ability to update bot's status on Discord
  - Ability to update bot's voice state on Discord
- - Ability to get roles from a guild using the messaging api
  - Support for re-sharding a bot as it grows
  - Docstrings for many namespaces
  - Support for tts in messages
+ - New HTTP endpoint support
+   - get-guild-roles
+   - create-guild-ban
+   - create-dm
+ - Logging when unable to recieve gateway information
  
 ### Changed
  - Moved many functions from `discljord.connections` to an implementation namespace
@@ -20,6 +24,7 @@ Discljord follows semantic versioning.
  
 ### Fixed
  - Improper handling of rate limits when none is found
+ - Bot no longer starts when out of identify packets for shards
  
 ### Depricated
  - `send-message!` function